(drive) ref: google drive modules + add: copy file action
